+++ src/readthedocs-config.js
@@ -1,5 +1,6 @@
 import { CLIENT_VERSION } from "./utils";
 import { EVENT_READTHEDOCS_ADDONS_DATA_READY } from "./events";
+import { ADDONS_API_VERSION } from "./utils";
 
 /**
  * Get the Read the Docs API version supported by user's integrations.
@@ -7,7 +8,7 @@ import { EVENT_READTHEDOCS_ADDONS_DATA_READY } from "./events";
  */
 function _getMetadataAddonsAPIVersion() {
   const meta = document.querySelector(
-    "meta[name=readthedocs-addons-api-version]"
+    "meta[name=readthedocs-addons-api-version]",
   );
   if (meta !== undefined) {
     return meta.getAttribute("content");
@@ -19,14 +20,34 @@ function _getMetadataAddonsAPIVersion() {
  * Load Read the Docs configuration from API endpoint.
  *
  */
-export function getReadTheDocsConfig() {
-  let url =
-    "/_/addons/?" +
-    new URLSearchParams({
-      url: window.location.href,
-      "client-version": CLIENT_VERSION,
-      "api-version": ADDONS_API_VERSION,
-    });
+export function getReadTheDocsConfig(sendUrlParam) {
+  const metaProject = document.querySelector(
+    "meta[name='readthedocs-project-slug']",
+  );
+  const metaVersion = document.querySelector(
+    "meta[name='readthedocs-version-slug']",
+  );
+
+  let projectSlug;
+  let versionSlug;
+  let params = {
+    "client-version": CLIENT_VERSION,
+    "api-version": ADDONS_API_VERSION,
+  };
+
+  if (sendUrlParam) {
+    params["url"] = window.location.href;
+  }
+
+  if (metaProject && metaVersion) {
+    projectSlug = metaProject.content;
+    versionSlug = metaVersion.content;
+
+    params["project-slug"] = projectSlug;
+    params["version-slug"] = versionSlug;
+  }
+
+  let url = "/_/addons/?" + new URLSearchParams(params);
 
   // Retrieve a static JSON file when working in development mode
   if (window.location.href.startsWith("http://localhost")) {
@@ -74,7 +95,7 @@ export function getReadTheDocsConfig() {
             .then((data) => {
               const readthedocsDataReady = new CustomEvent(
                 EVENT_READTHEDOCS_ADDONS_DATA_READY,
-                { detail: data }
+                { detail: data },
               );
               document.dispatchEvent(readthedocsDataReady);
               return undefined;
@@ -84,7 +105,7 @@ export function getReadTheDocsConfig() {
             EVENT_READTHEDOCS_ADDONS_DATA_READY,
             {
               detail: data,
-            }
+            },
           );
           document.dispatchEvent(readthedocsDataReady);
         }