+++ src/utils.js
@@ -1,6 +1,8 @@
+import { ajv } from "./data-validation";
+import { default as objectPath } from "object-path";
 export const ADDONS_API_VERSION = "0";
 // This is managed by bumpver automatically
-export const CLIENT_VERSION = "0.5.0";
+export const CLIENT_VERSION = "0.8.0";
 
 export function isReadTheDocsEmbedPresent() {
   const urls = [
@@ -33,7 +35,7 @@ export const domReady = new Promise((resolve) => {
         capture: true,
         once: true,
         passive: true,
-      }
+      },
     );
   }
 });
@@ -45,7 +47,32 @@ export const domReady = new Promise((resolve) => {
  * and loading.
  */
 export class AddonBase {
+  static jsonValidationURI = null;
+  static addonName = null;
+  static addonEnabledPath = null;
+
+  static isConfigValid(config) {
+    const validate = ajv.getSchema(this.jsonValidationURI);
+    const valid = validate(config);
+    if (!valid) {
+      console.debug(`Validation error on addon "${this.addonName}":`);
+      console.debug(validate.errors);
+    }
+    return valid;
+  }
+
   static isEnabled(config) {
+    return (
+      this.isConfigValid(config) &&
+      objectPath.get(config, this.addonEnabledPath, false) === true
+    );
+  }
+
+  static requiresUrlParam() {
+    // Decide whether or not this addons requires sending `url=` parameter to the API endpoint.
+    // Sending this attribute will make the API response to contain extra data (e.g. resolved URLs that depend on the exact URL)
+    //
+    // Note that sending `url=` attribute reduces the possibilities to use a cached response accross all the pages for the same project/version.
     return false;
   }
 }