+++ tests/docdiff.test.js
@@ -2,17 +2,7 @@ import { expect, assert, fixture, html } from "@open-wc/testing";
 import { DocDiffAddon, DocDiffElement } from "../src/docdiff";
 
 describe("Doc diff addon", () => {
-  it("is disabled by default", () => {
-    expect(
-      DocDiffAddon.isEnabled({
-        addons: {
-          doc_diff: {},
-        },
-      })
-    ).to.be.false;
-  });
-
-  it("is enabled with configuration", () => {
+  it("invalid configuration disables the addon", () => {
     expect(
       DocDiffAddon.isEnabled({
         addons: {
@@ -20,7 +10,33 @@ describe("Doc diff addon", () => {
             enabled: true,
           },
         },
-      })
+      }),
+    ).to.be.false;
+  });
+
+  it("is disabled with valid data", () => {
+    expect(
+      DocDiffAddon.isEnabled({
+        addons: {
+          doc_diff: {
+            enabled: false,
+            base_url: "http://project.readthedocs.io/en/latest/index.html",
+          },
+        },
+      }),
+    ).to.be.false;
+  });
+
+  it("is enabled with valid data", () => {
+    expect(
+      DocDiffAddon.isEnabled({
+        addons: {
+          doc_diff: {
+            enabled: true,
+            base_url: "http://project.readthedocs.io/en/latest/index.html",
+          },
+        },
+      }),
     ).to.be.true;
   });
 });