+++ tests/ethicalads.test.js
@@ -2,17 +2,7 @@ import { expect, assert, fixture, html } from "@open-wc/testing";
 import { EthicalAdsAddon } from "../src/ethicalads";
 
 describe("EthicalAds addon", () => {
-  it("is disabled by default", () => {
-    expect(
-      EthicalAdsAddon.isEnabled({
-        addons: {
-          ethicalads: {},
-        },
-      })
-    ).to.be.false;
-  });
-
-  it("is enabled with configuration", () => {
+  it("invalid configuration disables the addon", () => {
     expect(
       EthicalAdsAddon.isEnabled({
         addons: {
@@ -20,7 +10,39 @@ describe("EthicalAds addon", () => {
             enabled: true,
           },
         },
-      })
+      }),
+    ).to.be.false;
+  });
+
+  it("is disabled with valid data", () => {
+    expect(
+      EthicalAdsAddon.isEnabled({
+        addons: {
+          ethicalads: {
+            enabled: false,
+            ad_free: false,
+            campaign_types: ["community", "paid"],
+            keywords: ["docs", "data-science"],
+            publisher: "readthedocs",
+          },
+        },
+      }),
+    ).to.be.false;
+  });
+
+  it("is enabled with valid data", () => {
+    expect(
+      EthicalAdsAddon.isEnabled({
+        addons: {
+          ethicalads: {
+            enabled: true,
+            ad_free: false,
+            campaign_types: ["community", "paid"],
+            keywords: ["docs", "data-science"],
+            publisher: "readthedocs",
+          },
+        },
+      }),
     ).to.be.true;
   });
 
@@ -31,9 +53,12 @@ describe("EthicalAds addon", () => {
           ethicalads: {
             enabled: true,
             ad_free: true,
+            campaign_types: ["community", "paid"],
+            keywords: ["docs", "data-science"],
+            publisher: "readthedocs",
           },
         },
-      })
+      }),
     ).to.be.false;
   });
 });