+++ tests/flyout.test.js
@@ -2,17 +2,7 @@ import { expect, assert, fixture, html } from "@open-wc/testing";
 import { FlyoutAddon, FlyoutElement } from "../src/flyout";
 
 describe("Flyout addon", () => {
-  it("is disabled by default", () => {
-    expect(
-      FlyoutAddon.isEnabled({
-        addons: {
-          flyout: {},
-        },
-      })
-    ).to.be.false;
-  });
-
-  it("is enabled with configuration", () => {
+  it("invalid configuration disables the addon", () => {
     expect(
       FlyoutAddon.isEnabled({
         addons: {
@@ -20,7 +10,65 @@ describe("Flyout addon", () => {
             enabled: true,
           },
         },
-      })
+      }),
+    ).to.be.false;
+  });
+
+  it("is disabled with valid data", () => {
+    expect(
+      FlyoutAddon.isEnabled({
+        addons: {
+          flyout: {
+            enabled: false,
+            downloads: [],
+            translations: [],
+            versions: [],
+          },
+        },
+        domains: {
+          dashboard: "readthedocs.org",
+        },
+        projects: {
+          current: {
+            slug: "project",
+            single_version: false,
+          },
+        },
+        versions: {
+          current: {
+            slug: "latest",
+          },
+        },
+      }),
+    ).to.be.false;
+  });
+
+  it("is enabled with valid data", () => {
+    expect(
+      FlyoutAddon.isEnabled({
+        addons: {
+          flyout: {
+            enabled: true,
+            downloads: [],
+            translations: [],
+            versions: [],
+          },
+        },
+        domains: {
+          dashboard: "readthedocs.org",
+        },
+        projects: {
+          current: {
+            slug: "project",
+            single_version: false,
+          },
+        },
+        versions: {
+          current: {
+            slug: "latest",
+          },
+        },
+      }),
     ).to.be.true;
   });
 });