+++ tests/notification.test.js
@@ -4,14 +4,14 @@ import { NotificationElement } from "../src/notification";
 describe("Notification element", () => {
   it("is defined", async () => {
     const elem = await fixture(
-      html`<readthedocs-notification></readthedocs-notification>`
+      html`<readthedocs-notification></readthedocs-notification>`,
     );
     assert.instanceOf(elem, NotificationElement);
   });
 
   it("doesn't render if the config is missing", async () => {
     const elem = await fixture(
-      html`<readthedocs-notification></readthedocs-notification>`
+      html`<readthedocs-notification></readthedocs-notification>`,
     );
     assert.shadowDom.equal(elem, ``);
   });